+++ CHANGELOG.md
@@ -3,7 +3,11 @@
 All notable changes to the "Go-Web Framework" will be documented in this file.
 
 ## [Unreleased]
+### Added
 - Insert changelog
+
+### Fixed
+- Fix authentication flow in JWTAuth
 
 ## [v0.2-beta] - 2020-02-19
 ### Added